Consolidate upload utility helpers, remove from mvc path
